+++ terrareg/config.py
@@ -486,6 +486,18 @@ class Config:
         return os.environ.get('ADDITIONAL_MODULE_TABS', '[["Release Notes", ["RELEASE_NOTES.md", "CHANGELOG.md"]], ["License", ["LICENSE"]]]')
 
     @property
+    def ENABLE_ACCESS_CONTROLS(self):
+        """
+        Enables role based access controls for SSO users.
+
+        Enabling this feature will restrict all SSO users from performing any admin tasks.
+        Group mappings can be setup in the settings page, providing SSO users and groups with global or namespace-based permissions.
+
+        When disabled, all SSO users will have global admin privileges.
+        """
+        return self.convert_boolean(os.environ.get('ENABLE_ACCESS_CONTROLS', 'False'))
+
+    @property
     def OPENID_CONNECT_LOGIN_TEXT(self):
         """
         Text for sign in button for OpenID Connect authentication
@@ -514,6 +526,15 @@ class Config:
         A well-known URL will be expected at `${OPENID_CONNECT_ISSUER}/.well-known/openid-configuration`
         """
         return os.environ.get('OPENID_CONNECT_ISSUER', None)
+
+    @property
+    def OPENID_CONNECT_DEBUG(self):
+        """
+        Enable debug of OpenID connect via stdout.
+
+        This should only be enabled for non-production environments.
+        """
+        return self.convert_boolean(os.environ.get('OPENID_CONNECT_DEBUG', 'False'))
 
     @property
     def SAML2_LOGIN_TEXT(self):
@@ -569,6 +590,22 @@ class Config:
         ```
         """
         return os.environ.get('SAML2_PRIVATE_KEY', None)
+
+    @property
+    def SAML2_GROUP_ATTRIBUTE(self):
+        """
+        SAML2 user data group attribute.
+        """
+        return os.environ.get('SAML2_GROUP_ATTRIBUTE', 'groups')
+
+    @property
+    def SAML2_DEBUG(self):
+        """
+        Enable debug of Saml2 via stdout.
+
+        This should only be enabled for non-production environments.
+        """
+        return self.convert_boolean(os.environ.get('SAML2_DEBUG', 'False'))
 
     def convert_boolean(self, string):
         """Convert boolean environment variable to boolean."""